% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-dbl.R
\name{bits_to_dbl}
\alias{bits_to_dbl}
\title{Bits to double}
\usage{
bits_to_dbl(bits)
}
\arguments{
\item{bits}{vector of bits with 1 bit for sign, 11 bits for exponent and 52 bits for mantissa}
}
\value{
vector of bits, MSB first
}
\description{
Bits to double
}
