% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-int32.R
\name{bits_to_int32}
\alias{bits_to_int32}
\title{Convert 32 bits into a a single R 32 bit signed integer value}
\usage{
bits_to_int32(bits)
}
\arguments{
\item{bits}{vector of 32 bits, sign bit and then 31 bits with most-significant bit first.}
}
\value{
an integer value
}
\description{
Convert 32 bits into a a single R 32 bit signed integer value
}
