% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-int32.R
\name{bits_to_lofi}
\alias{bits_to_lofi}
\title{Convert a sequence of 0/1 values (MSB first) to an integer}
\usage{
bits_to_lofi(bits)
}
\arguments{
\item{bits}{vector of 0/1 values}
}
\value{
lofi
}
\description{
Convert a sequence of 0/1 values (MSB first) to an integer
}
