% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-choice.R
\name{choice_to_lofi}
\alias{choice_to_lofi}
\alias{lofi_to_choice}
\title{Convert a choice from a list of options to a lofi bit representation}
\usage{
choice_to_lofi(choice, options)

lofi_to_choice(lofi, options)
}
\arguments{
\item{choice}{value to convert to lofi representation}

\item{options}{the reference list of options it should be coded from}

\item{lofi}{low-bit representation}
}
\value{
low-bit representation i.e. the index of the choice in the list of options
}
\description{
Convert a choice from a list of options to a lofi bit representation
}
