% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-dbl.R
\name{dbl_to_bits}
\alias{dbl_to_bits}
\title{Convert a single double precision floating point value to bits}
\usage{
dbl_to_bits(dbl)
}
\arguments{
\item{dbl}{double value}
}
\value{
vector of bits, MSB first
}
\description{
Convert a single double precision floating point value to bits
}
