% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{expect_single_numeric}
\alias{expect_single_numeric}
\title{Expect the given value to be a single, non-NA numeric value}
\usage{
expect_single_numeric(value, value_desc = "value")
}
\arguments{
\item{value}{value}

\item{value_desc}{description of value to be inserted into error message}
}
\description{
Expect the given value to be a single, non-NA numeric value
}
