% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{expect_valid_nbits}
\alias{expect_valid_nbits}
\title{Validate 'nbits' to be a proper value}
\usage{
expect_valid_nbits(nbits, value_desc = "nbits")
}
\arguments{
\item{nbits}{number of bits}

\item{value_desc}{description of value to be inserted into error message}
}
\description{
Validate 'nbits' to be a proper value
}
