% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-dbl.R
\name{exponent_bits_to_int32}
\alias{exponent_bits_to_int32}
\title{Convert an exponent in offset binary format to a 32 bit signed integer}
\usage{
exponent_bits_to_int32(bits)
}
\arguments{
\item{bits}{vector of bits}
}
\value{
Return integer
}
\description{
Reference: \url{https://en.wikipedia.org/wiki/Half-precision_floating-point_format}
}
