% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-dbl.R
\name{exponent_range}
\alias{exponent_range}
\title{Calculate the possible range of exponents given the number of bits available}
\usage{
exponent_range(nbits_e)
}
\arguments{
\item{nbits_e}{Number of bits for the exponent}
}
\value{
Return 2-element vector giving the min and max possible exponent
}
\description{
Calculate the possible range of exponents given the number of bits available
}
