% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-dbl.R
\name{float_info}
\alias{float_info}
\title{What can the given float hold?}
\usage{
float_info(float_bits)
}
\arguments{
\item{float_bits}{3 element vector allocating bits to  sign, exponent and mantiassa}
}
\value{
named list of details about the values this float could represent
}
\description{
What can the given float hold?
}
