% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-hexcolour.R
\name{hex_colour_to_lofi}
\alias{hex_colour_to_lofi}
\alias{lofi_to_hex_colour}
\title{Quantize hex colours into a low-bit representation}
\usage{
hex_colour_to_lofi(cols, rgb_bits)

lofi_to_hex_colour(lofi, rgb_bits)
}
\arguments{
\item{cols}{character vector of hex colours e.g. \code{c('#ffaa70', '#00a0e2')}}

\item{rgb_bits}{3-element vector with number of bits to represent each of the
3 colour channels. Each element must be in the range [1,8]}

\item{lofi}{vector of integers. Each integer contains a low-bit representation
of a colour.}
}
\value{
Return integer vector. Each integer represents a colour.
}
\description{
Quantize a character vector of hex colours into a lofi (reduced bit) representation,
and output this as a vector of integers.
}
\details{
A hex colour (e.g. \code{'#ffaa70'}) is a character representation of a 24-bit colour i.e. 3 colour channels, each
represented by an 8 bit number.

\emph{lofi} colour is a representation of a colour using a low number of bits for
each colour channel, and then packed as an integer.

See wikipedia articles:
\itemize{
\item{\href{https://en.wikipedia.org/wiki/8-bit_color}{8-bit colour using 3-3-2 bits for RGB}}
\item{\href{https://en.wikipedia.org/wiki/List_of_monochrome_and_RGB_palettes}{Monochrome and RGB Palettes}}
\item{\href{https://en.wikipedia.org/wiki/List_of_color_palettes}{Colour Palettes}}
}
}
