% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-int32.R
\name{int32_to_bits}
\alias{int32_to_bits}
\title{Convert a single R 32-bit signed integer value to bits}
\usage{
int32_to_bits(int32)
}
\arguments{
\item{int32}{integer value}
}
\value{
vector of 32 bits, sign bit and then 31 bits with most-significant bit first.
}
\description{
Convert a single R 32-bit signed integer value to bits
}
