% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-dbl.R
\name{int32_to_exponent_bits}
\alias{int32_to_exponent_bits}
\title{Convert a 32 bit signed integer to a floating point exponent in offset binary format}
\usage{
int32_to_exponent_bits(int32, nbits)
}
\arguments{
\item{int32}{signed 32 bit integer}

\item{nbits}{number of bits for exponent}
}
\value{
Return integer
}
\description{
Reference: \url{https://en.wikipedia.org/wiki/Half-precision_floating-point_format}
}
