% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-int32.R
\name{int32_to_lofi}
\alias{int32_to_lofi}
\alias{lofi_to_int32}
\title{Convert a single R 32-bit signed integer value to a lower bit representation}
\usage{
int32_to_lofi(int32, nbits, signed = FALSE)

lofi_to_int32(lofi, nbits, signed = FALSE)
}
\arguments{
\item{int32}{a single R 32-bit signed integer value}

\item{nbits}{number of bits}

\item{signed}{boolean. default: FALSE}

\item{lofi}{integer value}
}
\value{
an integer value, only the lower 'nbits' contain valid information
}
\description{
Convert a single R 32-bit signed integer value to a lower bit representation
}
