% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-lgl.R
\name{lgl_to_lofi}
\alias{lgl_to_lofi}
\alias{lofi_to_lgl}
\title{Convert a logical to a lofi representation}
\usage{
lgl_to_lofi(lgl)

lofi_to_lgl(lofi)
}
\arguments{
\item{lgl}{single logical value}

\item{lofi}{lofi representation}
}
\value{
lofi representation
}
\description{
Convert a logical to a lofi representation
}
