% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-int32.R
\name{lofi_to_bits}
\alias{lofi_to_bits}
\title{Convert a low-bit integer into bit representation}
\usage{
lofi_to_bits(lofi, nbits)
}
\arguments{
\item{lofi}{integer}

\item{nbits}{number of bits}
}
\value{
vector of bits (length = nbits)
}
\description{
Limitations: max 32 bit signed integer
}
