% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofi-dbl.R
\name{lofi_to_fp}
\alias{lofi_to_fp}
\alias{dbl_to_fp}
\alias{fp_to_dbl}
\alias{fp_to_lofi}
\alias{change_format_fp}
\title{Unpack a floating point value into a list}
\usage{
lofi_to_fp(lofi, float_bits)

dbl_to_fp(dbl)

fp_to_dbl(fp)

fp_to_lofi(fp)

change_format_fp(fp, float_bits = c(1, 8, 7))
}
\arguments{
\item{lofi}{floating point number. stored as bits in a a single R integer}

\item{float_bits}{length (in number of bits) of sign, exponent and mantissa.}

\item{dbl}{64 bit R floating point real}

\item{fp}{named list}
}
\value{
Return a named list with elements sign, exponent and mantissa
}
\description{
Assumes LSB is in the 32 or 64 bit of the representation
}
