% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mask}
\alias{mask}
\title{Create an integer to bit-mask the lower 'nbits' of an integer}
\usage{
mask(nbits)
}
\arguments{
\item{nbits}{number of lower bits to mask}
}
\value{
integer
}
\description{
Create an integer to bit-mask the lower 'nbits' of an integer
}
