% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_custom}
\alias{pack_custom}
\alias{unpack_custom}
\alias{pack_integer}
\alias{unpack_integer}
\alias{pack_scaled}
\alias{unpack_scaled}
\alias{pack_logical}
\alias{unpack_logical}
\alias{pack_choice}
\alias{unpack_choice}
\alias{pack_double}
\alias{unpack_double}
\alias{pack_colour}
\alias{unpack_colour}
\title{Pack/unpack values of the given type}
\usage{
pack_custom(value, spec)

unpack_custom(value, spec)

pack_integer(value, spec)

unpack_integer(value, spec)

pack_scaled(value, spec)

unpack_scaled(value, spec)

pack_logical(value, spec)

unpack_logical(value, spec)

pack_choice(value, spec)

unpack_choice(value, spec)

pack_double(value, spec)

unpack_double(value, spec)

pack_colour(value, spec)

unpack_colour(value, spec)
}
\arguments{
\item{value}{value to pack}

\item{spec}{specification for encoding this value}
}
\description{
Pack/unpack values of the given type
}
