% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hexcolour.R
\name{plot_lofi_palette}
\alias{plot_lofi_palette}
\title{Display a vector of lofi colours.}
\usage{
plot_lofi_palette(lofi, rgb_bits, title = NULL)
}
\arguments{
\item{lofi}{integer vector of lofi colours}

\item{rgb_bits}{3-element vector with number of bits to represent each of the
3 colour channels. Each element must be in the range [1,8]}

\item{title}{title}
}
\description{
Display a vector of lofi colours.
}
