% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hexcolour.R
\name{plot_palette}
\alias{plot_palette}
\title{Display a vector of hex colours.}
\usage{
plot_palette(colours, ncol = 2, width = 40, height = 10,
  xfraction = 0.99, yfraction = 0.95)
}
\arguments{
\item{colours}{vector of hex colours e.g. \code{c('#ffffff', '#123456')}}

\item{ncol}{Number of columns}

\item{width, height}{dimensions of each colour swatch}

\item{xfraction, yfraction}{fill fraction of the swatch}
}
\description{
Display a vector of hex colours (given as character strings).
}
