% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hexcolour.R
\name{rgba_vec_to_hex_colour}
\alias{rgba_vec_to_hex_colour}
\alias{hex_colour_to_rgba_vec}
\alias{int32_to_hex_colour}
\alias{hex_colour_to_int32}
\alias{rgba_vec_to_int32}
\alias{int32_to_rgba_vec}
\title{Convert between equivalent 32-bit representations}
\usage{
rgba_vec_to_hex_colour(rgba_vec)

hex_colour_to_rgba_vec(hex_colour)

int32_to_hex_colour(int32)

hex_colour_to_int32(hex_colour)

rgba_vec_to_int32(rgba_vec)

int32_to_rgba_vec(int32)
}
\arguments{
\item{rgba_vec}{4-element RGBA colour with values in range [0, 255]}

\item{hex_colour}{string "#0a44f5', "#33ff12ff"}

\item{int32}{32 bit integer}
}
\value{
Hex colour with Alpha
}
\description{
Convert between equivalent 32-bit representations
}
