% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{sanitise_pack_spec}
\alias{sanitise_pack_spec}
\title{Sanitise a full pack_spec}
\usage{
sanitise_pack_spec(pack_spec)
}
\arguments{
\item{pack_spec}{pack_spec}
}
\value{
Return sanitised pack_spec
}
\description{
Sanitise a full pack_spec
}
