% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{sanitise_spec}
\alias{sanitise_spec}
\title{Sanitise an individual spec}
\usage{
sanitise_spec(spec)
}
\arguments{
\item{spec}{single spec}
}
\value{
Return sanitised spec
}
\description{
Sanitise an individual spec
}
