% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{sanitise_spec_choice}
\alias{sanitise_spec_choice}
\title{Sanitise a spec for an 'choice'}
\usage{
sanitise_spec_choice(spec)
}
\arguments{
\item{spec}{named list object with 'type', 'nbits', etc}
}
\value{
Return sanitised spec with any generic default values set
}
\description{
Sanitise a spec for an 'choice'
}
