% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{sanitise_spec_colour}
\alias{sanitise_spec_colour}
\title{Sanitise a spec for an 'colour'}
\usage{
sanitise_spec_colour(spec)
}
\arguments{
\item{spec}{named list object with 'type = "colour"', 'nbits' or 'rgb_bits'}
}
\value{
Return sanitised spec with any generic default values set
}
\description{
Sanitise a spec for an 'colour'
}
