% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{sanitise_spec_custom}
\alias{sanitise_spec_custom}
\title{Sanitise a spec for a 'custom' pack}
\usage{
sanitise_spec_custom(spec)
}
\arguments{
\item{spec}{named list object with 'type', 'nbits', etc}
}
\value{
Return sanitised spec with any generic default values set
}
\description{
Sanitise a spec for a 'custom' pack
}
