% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{sanitise_spec_double}
\alias{sanitise_spec_double}
\title{Sanitise a spec for a 'double'}
\usage{
sanitise_spec_double(spec)
}
\arguments{
\item{spec}{named list object with 'type', 'nbits', etc}
}
\value{
Return sanitised spec with any generic default values set
}
\description{
Sanitise a spec for a 'double'
}
