% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-spec-sanitise.R
\name{sanitise_spec_logical}
\alias{sanitise_spec_logical}
\title{Sanitise a spec for a 'logical'}
\usage{
sanitise_spec_logical(spec)
}
\arguments{
\item{spec}{named list object with 'type', 'nbits', etc}
}
\value{
Return sanitised spec with any generic default values set
}
\description{
Sanitise a spec for a 'logical'
}
