% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_encoded_value}
\alias{valid_encoded_value}
\title{Check a value fits in the given number of bits}
\usage{
valid_encoded_value(x, nbits)
}
\arguments{
\item{x}{value. integer, or something which can be corerced to integer}

\item{nbits}{number of bits in which to fit 'x'}
}
\description{
Check a value fits in the given number of bits
}
